port classic to UC18
This makes the classic dimension for UC18 available via the classic
snap. The approach is slightly different than what was done in UC16,
i.e. the base tarfile from ubuntu 18.04 is included in the snap
and that is used to prepare the chroot environment (see also
https://forum.snapcraft.io/t/7985 for a discussion about this).

Note that as part of the process the sandboxing of apt has to be
disabled currently.