create,classic: make the classic snap work on classic distributions
So far the classic snap was crashing in create because it made
assumptions about how /etc looks like and that users are managed
via extrausers.

This PR fixes this so that the classic snap will work on at least
regular Ubuntu releases. It's not *super* useful there but it may
be a convinient way to get a chroot with the right bind mounts
etc. To make it work the "classic" command will ensure the
SUDO_USER is available in the chroot and has sudo access in there.

Alternatively we could simply refuse to run in "classic.create"
if this is run on a non Ubuntu Core device.